Revert test suite and fix unit tests (#3)
The test class has been reintroduced to keep the force-di framework as
closely aligned to the origin as possible. This file is not ignored in
deployments to orgs rather than omitted from the repository.

Unit test fixtures that referenced a non-existent namespace have been
changed to remove the namespace now that the code takes namespaces into
account.

---------

Co-authored-by: Alastair Brown <alastair.browen@sage.com>